init user framework

+++ user/src/main/java/cn/hopever/platform/user/repository/ExampleTableRepository.java
@@ -6,6 +6,6 @@ import org.springframework.data.repository.CrudRepository;
 /**
  * Created by Donghui Huo on 2016/8/30.
  */
-public interface ExampleTableRepository extends CrudRepository<ExampleTable, Long> {
+public interface ExampleTableRepository extends CrudRepository<ExampleTable, Integer> {
     public ExampleTable findByUserName(String username);
 }
